+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.12.3] - 2018-02-23
+### Changed
+- prototype.creep.js, removed obsolete roles, bug fix
+- prototype.spawn.js, removed obsolete, refactored roles list
+- config.js, removed roles list, moved now in prototype.spawn.js
+
 ## [2.12.2] - 2018-02-23
 ### Changed
 - role.hero.js, refactored and bug fix